rename operator files
