+++ ewah/operators/base.py
