+++ ewah/operators/dynamodb.py
@@ -1,5 +1,5 @@
 from ewah.constants import EWAHConstants as EC
-from ewah.operators.base_operator import EWAHBaseOperator
+from ewah.operators.base import EWAHBaseOperator
 
 import boto3
